python/requirements.txt
Flask==2.2.2
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
WTForms==3.0.1
Jinja2==3.1.2
SQLAlchemy==2.0.0
typing_extensions==4.4.0
plaid_python==14.0.0
python-dotenv==0.15.0
itsdangerous==2.1.2
Werkzeug==2.2.2

click==8.1.3
MarkupSafe==2.1.2
zipp==3.12.0